requirements.txt
# Core
opencv-python>=4.8
numpy>=1.24
pyyaml>=6.0
pandas>=2.2
matplotlib>=3.8

# Detection (Ultralytics YOLOv8)
ultralytics>=8.0.200
torch>=2.0
torchvision

# FastAPI backend
fastapi>=0.110
uvicorn[standard]>=0.23

# Tracker
filterpy>=1.4    # for SORT’s Kalman filter

# Optional monitoring/logging
python-multipart>=0.0.6

# --- Tracking upgrades ---
deep-sort-realtime>=1.3.2
lapx>=0.5.7
scikit-learn>=1.4

# --- ONNX / TensorRT paths (choose one at runtime) ---
onnx>=1.16
onnxruntime>=1.18
# If you have NVIDIA GPU and want GPU ONNX:
onnxruntime-gpu>=1.18

# --- Backend persistence/auth ---
SQLAlchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
passlib[bcrypt]>=1.7
PyJWT>=2.9

bcrypt==4.0.1
passlib==1.7.4
python-jose[cryptography]
